# swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Books
  description: |
    Swagger for Books Api
    
    [Link to localhost:3000](http://localhost:3000/swagger.yaml)
    
    
    
  license:
    name: MIT
servers:
  - url: http://localhost:3000
paths:
  /books:
    get:
      summary: List all books or search by query parameter
      operationId: listBooks
      tags:
        - Books
      parameters:
        - in: query
          description: search for books by query parameter Title
          name: Title
          schema:
            type: string
      responses:
        '200':
          description: Array of books
          content:
            application/json:    
              schema:
                $ref: '#/components/schemas/ResponseBooks'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a book
      operationId: createBook
      tags:
      - Books
      requestBody:
        description: the book to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBook'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Book:
      properties:
        ISBN: 
          type: string
        Title: 
          type: string
        Author: 
          type: string
        Price: 
          type : number
        SellerEmail: 
          type: string
        Used: 
          type: boolean
        Location: 
          properties:
            City: 
              type: string
            Street: 
              type: string
    ResponseBook:
      allOf:
      - $ref: '#/components/schemas/Book'
      - properties:
          id:
            type: integer
          _id:
            type: string
            format: uuid
    ResponseBooks:
      type: array
      items:
        $ref: "#/components/schemas/ResponseBook"
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string